+++ src/js/app/store.js
@@ -1,7 +1,8 @@
-import { createStore } from 'redux';
+import { createStore, compose } from 'redux';
 import reducer from './reducer';
 import middleware, { runSagaMiddleware } from './middleware';
+import devToolsExtension from './dev-tools-extension';
 
-export default createStore(reducer, middleware);
+export default createStore(reducer, compose(middleware, devToolsExtension()));
 
 runSagaMiddleware();