+++ src/js/chuck-norris/index.js
@@ -1,3 +1,3 @@
 export reducer from './reducers';
-export saga from './saga';
+export saga from './sagas';
 export component from './components/ChuckNorris';