+++ src/js/todo-manager/components/AddTodo.js
@@ -1,6 +1,6 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
-import { addTodo } from '../actions/index';
+import { addTodo } from '../actions';
 import RaisedButton from 'material-ui/RaisedButton';
 import TextField from 'material-ui/TextField';
 import { Grid, Cell } from 'radium-grid';
@@ -15,13 +15,19 @@ class AddTodo extends Component {
   constructor(props) {
     super(props);
     this.state = { value: '', error: '' };
+    this.handleInputFocus = this.handleInputFocus.bind(this);
     this.handleInputChange = this.handleInputChange.bind(this);
     this.handleButtonClick = this.handleButtonClick.bind(this);
   }
 
   // on render, set focus on text field
   componentDidMount() {
-    this.refs.todoTextField.focus();
+    this.handleInputFocus();
+  }
+
+  // triggers focus on text field
+  handleInputFocus() {
+    this.todoTextField.focus();
   }
 
   // on input change, update state
@@ -33,10 +39,10 @@ class AddTodo extends Component {
   handleButtonClick() {
     if (this.state.value) {
       this.props.addTodo(this.state.value);
-      this.setState({ value: '', error: '' }, () => this.refs.todoTextField.focus());
+      this.setState({ value: '', error: '' }, () => this.handleInputFocus());
     }
     else {
-      this.setState({ error: 'Field is required' });
+      this.setState({ error: 'Field is required' }, () => this.handleInputFocus());
     }
   }
 
@@ -44,8 +50,12 @@ class AddTodo extends Component {
     return (
       <Grid>
         <Cell width="1/3">
+          {
+            // Using ref callback instead of string.
+            // See https://facebook.github.io/react/docs/more-about-refs.html
+          }
           <TextField
-            ref="todoTextField"
+            ref={ref => (this.todoTextField = ref)}
             hintText="Enter Todo..."
             errorText={this.state.error}
             value={this.state.value}