+++ src/js/todo-manager/components/FilterLink.js
@@ -1,17 +1,10 @@
 import { connect } from 'react-redux';
-import { setVisibilityFilter } from '../actions/index';
+import { setVisibilityFilter } from '../actions';
 import Link from './Link';
 
 const mapStateToProps = (state, ownProps) => ({
+  filter: ownProps.filter,
   active: ownProps.filter === state.visibilityFilter
 });
 
-const mapDispatchToProps = (dispatch, ownProps) => ({
-  onClick: () => {
-    dispatch(setVisibilityFilter(ownProps.filter));
-  }
-});
-
-const FilterLink = connect(mapStateToProps, mapDispatchToProps)(Link);
-
-export default FilterLink;
+export default connect(mapStateToProps, { setVisibilityFilter })(Link);