+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 0.4.0 - 2016-07-19
+
+* FEATURE - Enabled Redux Dev Tools.
+* FEATURE - `npm test ./test/abc` and `npm run test:watch ./test/abc` to run (and watch) only tests within `./test/abc`.
+* FEATURE - Configured `webpack-dev-server` to prevent "No 'Access-Control-Allow-Origin' header is present on the requested resource". 
+* FEATURE - Added `enzyme` and `es6-promise` dependencies.
+* REFACTOR - Ref Callback instead of Ref String. See `https://facebook.github.io/react/docs/more-about-refs.html`.
+* MAINTENANCE - Dependency updates.
+    * `history v3.0.0` still doesn't work with `react-router`. See https://github.com/reactjs/react-router/issues/3515
+* MAINTENANCE - Tested with `Node.js v6.2.2`.
+* BUG - `webpack-dev-server` not resolving Roboto font path in CSS file.
+* BUG - Lint both `src` and `test` dirs on `npm test` and `npm run ci`. Lint first before running tests.
+
 ## 0.3.3 - 2016-07-01
 
 * FEATURE - `npm run test:watch` - cross-platform approach to watch for changes in test files before rerunning the tests.